+++ examples/basic-mpa-webpack/src/pages/Dashboard/routes.ts
@@ -1,0 +1,3 @@
+import Index from './index';
+
+export default [{ path: '/', component: Index }];