+++ examples/basic-mpa-webpack/src/pages/Profile/store.ts
@@ -1,0 +1,6 @@
+import { createStore } from 'ice';
+import counter from './models/counter';
+
+const store = createStore({ counter });
+
+export default store;