+++ packages/build-mpa-config/src/index.ts
@@ -1,6 +1,8 @@
 import * as path from 'path';
 import { formatPath, checkExportDefaultDeclarationExists } from '@builder/app-helpers';
 import { IPluginAPI } from 'build-scripts';
+import type ReactGenerator from './generate/ReactGenerator';
+import type RaxGenerator from './generate/RaxGenerator';
 import generateEntry from './generate';
 import { FrameworkType, IGenerateResult } from './types';
 
@@ -16,11 +18,19 @@ interface IConfigOptions {
   framework?: FrameworkType;
   entries?: IEntries[];
   targetDir?: string;
+  executeGenerateTasks?: boolean;
 }
 
-export const generateMPAEntries = (api: IPluginAPI, options: IConfigOptions) => {
+interface MPAEntries extends IEntries {
+  generator?: RaxGenerator | ReactGenerator;
+  finalEntry: string;
+  runAppPath: string | null;
+  routesFilePath: string | null;
+}
+
+export const generateMPAEntries = (api: IPluginAPI, options: IConfigOptions): Record<string, MPAEntries> => {
   const { context } = api;
-  const { framework = 'rax', targetDir = '' } = options;
+  const { framework = 'rax', targetDir = '', executeGenerateTasks = true } = options;
   let { entries } = options;
   const { rootDir, commandArgs } = context;
   if (commandArgs.mpaEntry) {
@@ -39,15 +49,27 @@ export const generateMPAEntries = (api: IPluginAPI, options: IConfigOptions) => 
     let finalEntry = entryPath;
     let runAppPath = null;
     let routesFilePath;
+    let generator = null;
+    let generateTasks = [];
     if (isAppEntry || checkExportDefaultDeclarationExists(path.join(rootDir, 'src', source))) {
       const result = generateEntry(api, { framework, targetDir, pageEntry: entryPath, entryName, pageConfig, isAppEntry });
+      if (executeGenerateTasks) {
+        result.generateTasks.forEach((generateTask) => {
+          generateTask();
+        });
+      } else {
+        generateTasks = result.generateTasks;
+      }
       finalEntry = result.entryPath;
       runAppPath = result.runAppPath;
       routesFilePath = result.routesFilePath;
+      generator = result.generator;
     }
 
     parsedEntries[entryName] = {
       ...entry,
+      generator,
+      generateTasks,
       finalEntry,
       runAppPath,
       routesFilePath,
@@ -93,10 +115,14 @@ const setMPAConfig = (api, config, options: IConfigOptions) => {
   });
 
   if (config.plugins.has('document')) {
+    const filteredEntries = { ...parsedEntries };
+    // remove property which is unnecessary for plugin document
+    delete filteredEntries.generator;
+    delete filteredEntries.generateTasks;
     config.plugin('document').tap(args => {
       return [{
         ...args[0],
-        pages: Object.values(parsedEntries),
+        pages: Object.values(filteredEntries),
       }];
     });
   }