+++ packages/create-app-shared/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.2.3
+
+- [fix] revert AppProvider
+- [fix] `__pageConfig` types for rax MPA
+
 ## 1.2.2
 
 - [fix] miniapp withRouter